JN-1062 . import survey data updates
